ToPDF-PlugIn/requirements.txt
###############################################################################
# pip dependent packages
###############################################################################

###############################################################################
# default ARGOS-LABS SDK module
###############################################################################
alabs.common

###############################################################################
# For your dependent modules
###############################################################################
# ...

comtypes
img2pdf
pywin32